requirements.txt
pillow==10.0.1
pyperclip==1.8.0
requests==2.31.0
tkinterdnd2==0.4.2
